+++ t-rex-service/src/mvt_service.rs
@@ -457,10 +457,8 @@ impl MvtService {
         cfg.push_str(&layer.gen_runtime_config());
         if let &Datasource::Postgis(ref pg) = ds {
             if layer.query(0).is_none() {
-                let zoom = 0; // relevant?
-                let query = pg.build_query_sql(layer, 3857, zoom, None, true).unwrap();
-                // Remove quotes from column names for better readability
-                cfg.push_str(&format!("#sql = \"\"\"{}\"\"\"\n", query.replace('"', "")))
+                let query = pg.build_query_sql_template(layer);
+                cfg.push_str(&format!("#sql = \"\"\"{}\"\"\"\n", query))
             }
         }
         cfg
@@ -478,10 +476,10 @@ impl<'a> Config<'a, ApplicationCfg> for MvtService {
             .collect();
         let cache = Tilecache::from_config(&config)?;
         Ok(MvtService {
-            datasources: datasources,
-            grid: grid,
-            tilesets: tilesets,
-            cache: cache,
+            datasources,
+            grid,
+            tilesets,
+            cache,
         })
     }
     fn gen_config() -> String {